// src/app/api/fn/documents/list-documents.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { DocumentList } from '../../models/document-list';
import { DocumentType } from '../../models/document-type';

export interface ListDocuments$Params {

/**
 * Filter by document type.
 */
  type?: DocumentType;

/**
 * Filter to documents created from this customer (works even after the customer is soft-deleted).
 */
  customer_id?: string;
  page?: number;
  per_page?: number;
}

export function listDocuments(http: HttpClient, rootUrl: string, params?: ListDocuments$Params, context?: HttpContext): Observable<StrictHttpResponse<DocumentList>> {
  const rb = new RequestBuilder(rootUrl, listDocuments.PATH, 'get');
  if (params) {
    rb.query('type', params.type, {});
    rb.query('customer_id', params.customer_id, {});
    rb.query('page', params.page, {});
    rb.query('per_page', params.per_page, {});
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'application/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<DocumentList>;
    })
  );
}

listDocuments.PATH = '/documents';